refactor(cs): introduce constants.js; narrow config.js to runtime config
Split gemini-tracker's static identifiers across two files with clear semantic boundaries:

- config.js: retains only true runtime config (STORAGE_KEY, URLs, SCHEMA_VERSION)
- constants.js (new): domain enums extracted from config.js (MODEL_NAMES, TOOL_NAMES)
  plus new named constants to replace inline magic strings:
    BROWSER_ACTIONS.UPDATE_HISTORY_COUNT — replaces hardcoded message action string
    GEMINI_PLANS.PRO / .FREE — replaces bare plan string literals in model-detector
    ERROR_PATTERNS — replaces inline array in crash-detector.isErrorMessage()
    STATUS_TYPES — replaces repeated status string literals in status-indicator;
      Object.values(STATUS_TYPES) now drives the classList.remove calls,
      so adding a new type only requires one edit in constants.js
